+++ SampleTest.Tests/TestsExplicit.cs
@@ -1,12 +1,14 @@
 ﻿using NUnit.Framework;
 using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace SampleTest.Tests
 {
     [Explicit]
-    public class TestsExplicit1_ShouldFail
+    public class TestsExplicit1_ShouldFail_Constructor
     {
-        public TestsExplicit1_ShouldFail()
+        public TestsExplicit1_ShouldFail_Constructor()
         {
             throw new NotImplementedException();
         }
@@ -19,7 +21,7 @@ namespace SampleTest.Tests
     }
 
     [Explicit]
-    public class TestsExplicit2_ShouldFail : IDisposable
+    public class TestsExplicit2_ShouldFail_Dispose : IDisposable
     {
         [Test]
         public void TestDispose()
@@ -34,7 +36,7 @@ namespace SampleTest.Tests
     }
 
     [Explicit]
-    public class TestsExplicit3_ShouldFail
+    public class TestsExplicit3_ShouldFail_OneTimeSetUp
     {
         [OneTimeSetUp]
         public void OneTimeSetUp()
@@ -50,7 +52,7 @@ namespace SampleTest.Tests
     }
 
     [Explicit]
-    public class TestsExplicit4_ShouldFail
+    public class TestsExplicit4_ShouldFail_OneTimeTearDown
     {
         [OneTimeTearDown]
         public void OneTimeTearDown()
@@ -66,7 +68,7 @@ namespace SampleTest.Tests
     }
 
     [Explicit]
-    public class TestsExplicit5_ShouldFail
+    public class TestsExplicit5_ShouldFail_SetUp
     {
         [SetUp]
         public void SetUp()
@@ -82,7 +84,7 @@ namespace SampleTest.Tests
     }
 
     [Explicit]
-    public class TestsExplicit6_ShouldFail
+    public class TestsExplicit6_ShouldFail_TearDown
     {
         [TearDown]
         public void TearDown()
@@ -98,21 +100,67 @@ namespace SampleTest.Tests
     }
 
     [Explicit]
-    public class TestsExplicit_ShouldFail
+    public class TestsExplicit_ShouldFail_Parameter
     {
         [Test]
-        public void TestParameterText(string text)
+        public void TestParameterText(string value)
         {
-            Console.WriteLine($"Value is {text}");
-            Assert.IsNotNull(text);
-            Assert.IsNotEmpty(text);
+            Console.WriteLine($"Value is {value}");
+            Assert.IsNotNull(value);
+            Assert.IsNotEmpty(value);
+            Assert.Pass($"{value} {value.GetType()}");
         }
 
         [Test]
-        public void TestParameterInteger(int integer)
+        public void TestParameterInteger(int value)
         {
-            Console.WriteLine($"Value is {integer}");
-            Assert.IsTrue(integer > 0);
+            Console.WriteLine($"Value is {value}");
+            Assert.IsTrue(value > 0);
+            Assert.Pass($"{value} {value.GetType()}");
+        }
+
+        [Test]
+        public void TestParameterLong(long value)
+        {
+            Console.WriteLine($"Value is {value}");
+            Assert.IsTrue(value > 0);
+            Assert.Pass($"{value} {value.GetType()}");
+        }
+
+        [Test]
+        public void TestParameterMultiple(int value, long value2)
+        {
+            Console.WriteLine($"Value is {value}");
+            Console.WriteLine($"Value2 is {value2}");
+            Assert.IsTrue(value > 0);
+            Assert.IsTrue(value2 > 0);
+            Assert.Pass($"{value} {value2}");
+        }
+
+        [Test]
+        public void TestParameterArray(string[] value)
+        {
+            Console.WriteLine($"Value is {value}");
+            Assert.IsNotNull(value);
+            Assert.IsNotEmpty(value);
+            Assert.Pass($"{value} {value.GetType()}");
+        }
+
+        [Test]
+        public void TestParameterIEnumerable(IEnumerable<string> value)
+        {
+            Console.WriteLine($"Value is {value}");
+            Assert.IsNotNull(value);
+            Assert.IsNotEmpty(value);
+            Assert.Pass($"{value} {value.GetType()}");
+        }
+
+        [Test]
+        public void TestParameterInterface(ICloneable value)
+        {
+            Console.WriteLine($"Value is {value}");
+            Assert.IsNotNull(value);
+            Assert.Pass($"{value} {value.GetType()}");
         }
     }
 }