+++ ricaun.NUnit/Services/TestAttributeService.cs
@@ -23,55 +23,160 @@ namespace ricaun.NUnit.Services
         }
 
         /// <summary>
-        /// GetFilterTestMethods
+        /// AnyMethodWithTestAttribute
         /// </summary>
         /// <param name="type"></param>
         /// <returns></returns>
-        public IEnumerable<MethodInfo> GetFilterTestMethods(Type type)
+        public bool AnyMethodWithTestAttribute(Type type)
         {
-            var methods = type.GetMethods().OrderBy(e => e.Name);
-            var testMethods = methods.Where(AnyTestAttribute);
-            var onlyTypeTests = testMethods.Where(m => GetTestAttributes(m).Any(a => HasFilterTestMethod(m, a)));
-            //foreach (var testMethod in testMethods)
-            //{
-            //    //Debug.WriteLine($"{testMethod.Name} {testMethod.GetCustomAttributes(true).Count()}");
-            //    Debug.WriteLine($"{testMethod.Name} {string.Join(" ", testMethod.GetCustomAttributes(true).Select(e => e.GetType().Name))}");
-            //    Debug.WriteLine($"{testMethod.Name} {testMethod.GetCustomAttributes(true).OfType<TestAttribute>().Count()}");
-            //}
-            //Debug.WriteLine($"Debug:\t{type}:\t{string.Join(" ", onlyTypeTests.Select(e => e.Name))}");
-            return onlyTypeTests;
+            return GetMethodWithTestAttribute(type).Any();
         }
+
+        /// <summary>
+        /// GetMethodWithTestAttribute
+        /// </summary>
+        /// <param name="type"></param>
+        /// <returns></returns>
+        public IEnumerable<MethodInfo> GetMethodWithTestAttribute(Type type)
+        {
+            return type.GetMethods().Where(AnyTestAttribute).OrderBy(e => e.Name);
+        }
+
+        /// <summary>
+        /// AnyMethodWithTestAttributeAndFilter
+        /// </summary>
+        /// <param name="type"></param>
+        /// <returns></returns>
+        public bool AnyMethodWithTestAttributeAndFilter(Type type)
+        {
+            return GetMethodWithTestAttributeAndFilter(type).Any();
+        }
+
+        /// <summary>
+        /// GetMethodWithTestAttributeAndFilter
+        /// </summary>
+        /// <param name="type"></param>
+        /// <returns></returns>
+        public IEnumerable<MethodInfo> GetMethodWithTestAttributeAndFilter(Type type)
+        {
+            return GetMethodWithTestAttribute(type).Where(m => GetTestAttributes(m).Any(a => HasFilterTestMethod(type, m, a)));
+        }
+
+        ///// <summary>
+        ///// GetFilterTestMethods
+        ///// </summary>
+        ///// <param name="type"></param>
+        ///// <returns></returns>
+        //[Obsolete]
+        //public IEnumerable<MethodInfo> GetFilterTestMethods(Type type)
+        //{
+        //    var methods = type.GetMethods().OrderBy(e => e.Name);
+        //    var testMethods = methods.Where(AnyTestAttribute);
+        //    var onlyTypeTests = testMethods.Where(m => GetTestAttributes(m).Any(a => HasFilterTestMethod(m, a)));
+        //    //foreach (var testMethod in testMethods)
+        //    //{
+        //    //    //Debug.WriteLine($"{testMethod.Name} {testMethod.GetCustomAttributes(true).Count()}");
+        //    //    Debug.WriteLine($"{testMethod.Name} {string.Join(" ", testMethod.GetCustomAttributes(true).Select(e => e.GetType().Name))}");
+        //    //    Debug.WriteLine($"{testMethod.Name} {testMethod.GetCustomAttributes(true).OfType<TestAttribute>().Count()}");
+        //    //}
+        //    //Debug.WriteLine($"Debug:\t{type}:\t{string.Join(" ", onlyTypeTests.Select(e => e.Name))}");
+        //    return onlyTypeTests;
+        //}
+
+        ///// <summary>
+        ///// HasFilterTestMethod
+        ///// </summary>
+        ///// <param name="method"></param>
+        ///// <param name="nUnitAttribute"></param>
+        ///// <returns></returns>
+        //public bool HasFilterTestMethod(MethodInfo method, NUnitAttribute nUnitAttribute)
+        //{
+        //    return TestEngineFilter.HasName(GetTestFullName(method, nUnitAttribute));
+        //}
+        //public bool HasFilterTestMethod(string typeFullName, MethodInfo method, NUnitAttribute nUnitAttribute)
+        //{
+        //    return TestEngineFilter.HasName(GetTestFullName(typeFullName, method, nUnitAttribute));
+        //}
 
         /// <summary>
         /// HasFilterTestMethod
         /// </summary>
+        /// <param name="type"></param>
         /// <param name="method"></param>
         /// <param name="nUnitAttribute"></param>
         /// <returns></returns>
-        public bool HasFilterTestMethod(MethodInfo method, NUnitAttribute nUnitAttribute)
+        public bool HasFilterTestMethod(Type type, MethodInfo method, NUnitAttribute nUnitAttribute)
         {
-            return TestEngineFilter.HasName(GetTestFullName(method, nUnitAttribute));
+            return TestEngineFilter.HasName(GetTestFullName(type, method, nUnitAttribute));
+        }
+
+        ///// <summary>
+        ///// GetMethodFullName
+        ///// </summary>
+        ///// <param name="method"></param>
+        ///// <returns></returns>
+        //[Obsolete]
+        //public string GetMethodFullName(MethodInfo method)
+        //{
+        //    return GetMethodFullName(method.DeclaringType, method);
+        //}
+
+        /// <summary>
+        /// GetMethodFullName
+        /// </summary>
+        /// <param name="declaringType"></param>
+        /// <param name="method"></param>
+        /// <returns></returns>
+        public string GetMethodFullName(Type declaringType, MethodInfo method)
+        {
+            return GetMethodFullName(declaringType.FullName, method);
         }
 
         /// <summary>
         /// GetMethodFullName
         /// </summary>
+        /// <param name="fullNameType"></param>
         /// <param name="method"></param>
         /// <returns></returns>
-        public string GetMethodFullName(MethodInfo method)
+        public string GetMethodFullName(string fullNameType, MethodInfo method)
         {
-            return method.DeclaringType.FullName + "." + method.Name;
+            return fullNameType + "." + method.Name;
+        }
+
+        ///// <summary>
+        ///// GetTestFullName
+        ///// </summary>
+        ///// <param name="method"></param>
+        ///// <param name="nUnitAttribute"></param>
+        ///// <returns></returns>
+        //[Obsolete]
+        //public string GetTestFullName(MethodInfo method, NUnitAttribute nUnitAttribute)
+        //{
+        //    return GetMethodFullName(method) + "." + GetTestName(method, nUnitAttribute);
+        //}
+
+        /// <summary>
+        /// GetTestFullName
+        /// </summary>
+        /// <param name="type"></param>
+        /// <param name="method"></param>
+        /// <param name="nUnitAttribute"></param>
+        /// <returns></returns>
+        public string GetTestFullName(Type type, MethodInfo method, NUnitAttribute nUnitAttribute)
+        {
+            return GetMethodFullName(type, method) + "." + GetTestName(method, nUnitAttribute);
         }
 
         /// <summary>
         /// GetTestFullName
         /// </summary>
+        /// <param name="fullNameType"></param>
         /// <param name="method"></param>
         /// <param name="nUnitAttribute"></param>
         /// <returns></returns>
-        public string GetTestFullName(MethodInfo method, NUnitAttribute nUnitAttribute)
+        public string GetTestFullName(string fullNameType, MethodInfo method, NUnitAttribute nUnitAttribute)
         {
-            return GetMethodFullName(method) + "." + GetTestName(method, nUnitAttribute);
+            return GetMethodFullName(fullNameType, method) + "." + GetTestName(method, nUnitAttribute);
         }
 
         /// <summary>
@@ -97,6 +202,13 @@ namespace ricaun.NUnit.Services
                 var name = testCaseAttribute.TestName ?? $"{method.Name}({string.Join(",", testCaseAttribute.Arguments)})";
                 return name;
             }
+
+            var parameters = method.GetParameters();
+            if (parameters.Any())
+            {
+                return $"{method.Name}({string.Join(",", parameters.Select(e => e.ParameterType.Name))})";
+            }
+
             return method.Name;
         }